getReceiptOracle return proof
